Fix #6339.
Former-commit-id: 802059a29fba0aab9c304c335b6be01413caccde

+++ ftp/src/main/java/ch/cyberduck/core/ftp/FTPDeleteFeature.java
@@ -25,9 +25,6 @@ import ch.cyberduck.core.features.Delete;
 import java.io.IOException;
 import java.util.List;
 
-/**
- * @version $Id$
- */
 public class FTPDeleteFeature implements Delete {
 
     private FTPSession session;
@@ -47,6 +44,10 @@ public class FTPDeleteFeature implements Delete {
                     }
                 }
                 else if(file.isDirectory()) {
+                    // Change working directory to parent
+                    if(!session.getClient().changeWorkingDirectory(file.getParent().getAbsolute())) {
+                        throw new FTPException(session.getClient().getReplyCode(), session.getClient().getReplyString());
+                    }
                     if(!session.getClient().removeDirectory(file.getAbsolute())) {
                         throw new FTPException(session.getClient().getReplyCode(), session.getClient().getReplyString());
                     }
